LayoutTests/imported/w3c/web-platform-tests/content-security-policy/inheritance/blob-url-inherits-from-initiator.sub-expected.txt


FAIL Initiator is same-origin with target frame. assert_equals: Eval should be blocked by CSP in blob URL. expected "eval blocked" but got "eval allowed"
FAIL Initiator is cross-origin with target frame. assert_equals: Eval should be blocked by CSP in blob URL. expected "eval blocked" but got "eval allowed"
